DONE LAST DIGIT

+++ 0x02-functions_nested_loops/7-print_last_digit.c
@@ -9,7 +9,7 @@
 
 int print_last_digit(int n)
 {
-	int rem;
+	long long rem;
 
 	if (n > 0)
 		rem = n % 10;
